api/requirements.txt
fastapi==0.110.3
uvicorn==0.29.0
GitPython==3.1.43
sentence-transformers==2.7.0
qdrant-client==1.9.0
python-dotenv==1.0.1
toml==0.10.2
openai
langchain
